Corrigir erros de tipo adicionais
Correções implementadas:
- Importar runes_data.dart em rune_spread_model.dart para ter acesso ao typedef RuneModel
- Converter degree de int para double em transit_calculator.dart (position.degree.toDouble())
- Converter natalPlanet.degree para double (natalPlanet.degree.toDouble())
- Corrigir _parseMoonPhase para retornar MoonPhase? em vez de String?
  usando MoonPhase.values.firstWhere para encontrar o enum correto

Todos os erros de compilação restantes foram corrigidos.

+++ lib/features/astrology/data/services/transit_calculator.dart
@@ -20,7 +20,7 @@ class TransitCalculator {
       transits.add(Transit(
         planet: planet,
         sign: position.sign,
-        degree: position.degree,
+        degree: position.degree.toDouble(),
         isRetrograde: position.isRetrograde,
       ));
     }
@@ -43,7 +43,7 @@ class TransitCalculator {
       // Verificar aspectos com planetas natais
       for (final natalPlanet in natalChart.planets) {
         final natalLongitude =
-            _getAbsoluteLongitude(natalPlanet.sign, natalPlanet.degree);
+            _getAbsoluteLongitude(natalPlanet.sign, natalPlanet.degree.toDouble());
 
         // Calcular diferença angular
         var diff = (transitLongitude - natalLongitude).abs();
